Transition to CodeAuroraForum Camera
Unlike frameworks, which allowed a squashed cherry-pick of the
necessary items to bring the camera current, this merge required
multiple alterations to maintain the existing features. Power key
shutter, panorama functionality, and preview override are not available
through CAF and caused multiple instances of merge failure. The most
efficient method was to perform a single commit of the necessary
changes to provide the best of both options.